+++ src/main/java/dansplugins/mailboxes/commands/ConfigCommand.java
@@ -57,7 +57,7 @@ public class ConfigCommand {
         }
         else if (args[0].equalsIgnoreCase("set")) {
             if (args.length < 3) {
-                sender.sendMessage(ChatColor.RED + "Usage: /c config set (option) (value)");
+                sender.sendMessage(ChatColor.RED + "Usage: /m config set (option) (value)");
                 return false;
             }
             String option = args[1];